% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fit.R
\name{extract_fit}
\alias{extract_fit}
\title{Extract fit statistics}
\usage{
extract_fit(..., details = FALSE)
}
\arguments{
\item{...}{lavaan objects}

\item{details}{Logical, specify whether to return detailed fir statistics}
}
\value{

}
\description{
Extract fit statistics
}
