% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_minvariance.R
\name{long_minvariance}
\alias{long_minvariance}
\title{Test longitudinal measurement invariance}
\usage{
long_minvariance(
  data,
  n_timepoints,
  n_items,
  add = NULL,
  remove = NULL,
  measure_name = "x",
  time_str = "_t",
  item_str = "_i",
  details = FALSE,
  return = c("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax")
)
}
\arguments{
\item{data}{Data in wide format (one row for each participant, one column for each item at each time point)}

\item{n_timepoints}{Number of time points}

\item{n_items}{Number of items}

\item{add}{lavaan syntax to add to the model}

\item{remove}{List, remove specific parts of the automatically generated model}

\item{measure_name}{String, name of the measure}

\item{time_str}{String, prefix of the point specifier}

\item{item_str}{String, prefix of the item specifier}

\item{details}{Return details of selected output selected in return argument}

\item{return}{What should be returned ("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax")}
}
\value{

}
\description{
Test longitudinal measurement invariance
}
