% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_lavaan_syntax.R
\name{long_minvariance_syntax}
\alias{long_minvariance_syntax}
\title{Title}
\usage{
long_minvariance_syntax(
  n_items,
  n_timepoints,
  add = NULL,
  model = c("configural", "weak", "strong", "strict"),
  measure_name = "x",
  time_str = "_t",
  item_str = "_i"
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
