% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_long_minvariance.R
\name{minvariance}
\alias{minvariance}
\title{Test longitudinal measurement invariance}
\usage{
minvariance(
  data,
  var_list,
  add = NULL,
  remove = NULL,
  details = FALSE,
  return = c("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax", "all")
)
}
\arguments{
\item{data}{Data in wide format (one row for each participant, one column for each item at each time point)}

\item{var_list}{List specifying the variable names for each time point. The number of variable names need to be identical at each time point (same measure).}

\item{add}{String, lavaan syntax to be added to the model}

\item{remove}{List, remove specific parts of the automatically generated model}

\item{details}{Return details of selected output selected in return argument}

\item{return}{What should be returned ("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax", "all")}

\item{measure_name}{String, name of the measure}

\item{time_str}{String, prefix of the point specifier}

\item{item_str}{String, prefix of the item specifier}
}
\value{

}
\description{
This function automatically runs 4 automatically generated measurement invariance models.
This may not be appropriate and it might be more useful to generate lavaan syntax using \link[minvariance]{long_minvariance_syntax} function for each model separately and run confirmatory factor analyses using the \link[lavaan]{cfa} function from lavaan.
}
