% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specyfy_lavaan_syntax.R
\name{specify_latent_factors}
\alias{specify_latent_factors}
\title{Specify lavaan syntax for latent factors}
\usage{
specify_latent_factors(
  n_items,
  measure_name,
  n_timepoints,
  model = c("configural", "weak", "strong", "strict")
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Specify lavaan syntax for latent factors
}
