% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specyfy_lavaan_syntax.R
\name{specify_latent_variable_covariances}
\alias{specify_latent_variable_covariances}
\title{Title}
\usage{
specify_latent_variable_covariances(
  n_items,
  n_timepoints,
  model = c("configural", "weak", "strong", "strict"),
  measure_name = "x",
  time_str = "s",
  item_str = "i"
)
}
\arguments{
\item{item_str}{}
}
\value{

}
\description{
Title
}
