% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specyfy_lavaan_syntax.R
\name{specify_latent_variable_means}
\alias{specify_latent_variable_means}
\title{Title}
\usage{
specify_latent_variable_means(
  n_items,
  measure_name,
  n_timepoints,
  model = c("configural", "weak", "strong", "strict")
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
