% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_minvariance.R
\name{long_minvariance}
\alias{long_minvariance}
\title{Title}
\usage{
long_minvariance(
  data,
  n_timepoints,
  n_items,
  measure_name = "x",
  details = FALSE,
  return = c("fit_statistics", "model_tests", "lavaan_objects", "lavaan_syntax")
)
}
\arguments{
\item{return}{}
}
\value{

}
\description{
Title
}
