% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specyfy_lavaan_syntax.R
\name{long_minvariance_syntax}
\alias{long_minvariance_syntax}
\title{Title}
\usage{
long_minvariance_syntax(
  n_items,
  measure_name,
  n_timepoints,
  add = NULL,
  model = c("configural", "weak", "strong", "strict")
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
