% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_lavaan_syntax.R
\name{specify_latent_variable_means}
\alias{specify_latent_variable_means}
\title{Title}
\usage{
specify_latent_variable_means(
  n_items,
  n_timepoints,
  model = c("configural", "weak", "strong", "strict"),
  measure_name = "x",
  time_str = "_t",
  item_str = "_i"
)
}
\arguments{
\item{model}{}
}
\value{

}
\description{
Title
}
